+++ system/lib/wasmfs/backends/noderawfs_backend.cpp
@@ -3,8 +3,6 @@
 // University of Illinois/NCSA Open Source License.  Both these licenses can be
 // found in the LICENSE file.
 
-#include <memory>
-
 #include "emscripten/wasmfs.h"
 
 namespace wasmfs {
